// class-room-sauvegarde/11/class-room-frontend/src/app/api/model/periodeDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EtablissementProfesseur } from './etablissementProfesseur';


export interface PeriodeDto { 
    idPeriode?: number;
    libelle?: string;
    dateDebut?: string;
    dateFin?: string;
    etablissementProfesseur: EtablissementProfesseur;
}